changed double to float, separated public/private interface

+++ ktf2cpp/codegen/lib/profile_predictor/main.c
@@ -4,8 +4,8 @@ int main()
 {
   /* printf("hello world"); */
   /*  printf("%f",fmax(1.0,0.0)); */
-  double inputs[256] = {0};
-  double outputs[30] = {0};
+  float inputs[256] = {0};
+  float outputs[30] = {0};
   profile_predictor(inputs, outputs);
     for (int i=0 ; i<30 ; i++)
       printf("%f\n", outputs[i]);
